+++ hydra-node/test/Hydra/Chain/Direct/Contract/Close.hs
@@ -36,7 +36,7 @@ import Hydra.Data.ContestationPeriod (posixFromUTCTime)
 import qualified Hydra.Data.ContestationPeriod as OnChain
 import qualified Hydra.Data.Party as OnChain
 import Hydra.Ledger (hashUTxO)
-import Hydra.Ledger.Cardano (genOneUTxOFor, genVerificationKey)
+import Hydra.Ledger.Cardano (genOneUTxOFor, genValue, genVerificationKey)
 import Hydra.Ledger.Cardano.Evaluate (genValidityBoundsFromContestationPeriod)
 import Hydra.Party (Party, deriveParty, partyToChain)
 import Hydra.Snapshot (Snapshot (..), SnapshotNumber)
@@ -223,6 +223,8 @@ data CloseMutation
     MutateTokenMintingOrBurning
   | -- | Change the resulting contesters to non-empty to see if they are checked
     MutateContesters
+  | -- | See spec: 5.5. rule 6 -> value is preserved
+    MutateValueInOutput
   deriving (Generic, Show, Enum, Bounded)
 
 genCloseMutation :: (Tx, UTxO) -> Gen SomeMutation
@@ -284,6 +286,9 @@ genCloseMutation (tx, _utxo) =
     , SomeMutation (Just "contesters non-empty") MutateContesters . ChangeOutput 0 <$> do
         mutatedContesters <- listOf1 $ PubKeyHash . toBuiltin <$> genHash
         pure $ headTxOut & changeHeadOutputDatum (replaceContesters mutatedContesters)
+    , SomeMutation (Just "head value is not preserved") MutateValueInOutput <$> do
+        newValue <- genValue
+        pure $ ChangeOutput 0 (headTxOut{txOutValue = newValue})
     ]
  where
   genOversizedTransactionValidity = do