+++ hydra-plutus/src/Hydra/Contract/Head.hs
@@ -182,7 +182,7 @@ checkCollectCom ctx@ScriptContext{scriptContextTxInfo = txInfo} (contestationPer
     && mustNotChangeParameters
     && everyoneHasCommitted
     && mustBeSignedByParticipant ctx headId
-    && hasST headId outValue
+    && hasST headId val
  where
   mustCollectUtxoHash =
     traceIfFalse "incorrect utxo hash" $
@@ -208,7 +208,7 @@ checkCollectCom ctx@ScriptContext{scriptContextTxInfo = txInfo} (contestationPer
       _ -> traceError "wrong state in output datum"
   headAddress = mkHeadAddress ctx
 
-  outValue =
+  val =
     maybe mempty (txOutValue . txInInfoResolved) $ findOwnInput ctx
 
   everyoneHasCommitted =
@@ -261,6 +261,8 @@ commitDatum txInfo input = do
 --   * State token (ST) is present in the output
 --
 --   * Contesters must be initialize as empty
+--
+--   * Value in v_head is preserved
 checkClose ::
   ScriptContext ->
   [Party] ->
@@ -275,16 +277,26 @@ checkClose ctx parties initialUtxoHash sig cperiod headPolicyId =
     && checkDeadline
     && checkSnapshot
     && mustBeSignedByParticipant ctx headPolicyId
-    && hasST headPolicyId outValue
     && mustInitializeContesters
+    && hasST headPolicyId val
+    && mustPreserveValue
     && mustNotChangeParameters
  where
+  mustPreserveValue =
+    traceIfFalse "head value is not preserved" $
+      -- XXX: Equality on value is very memory intensive as it's defined on
+      -- associative lists and Map equality is implemented. Instead we should be
+      -- more strict and require EXACTLY the same value and compare using a
+      -- simple fold or even compare the serialised bytes.
+      val == val'
+
+  val' = txOutValue . head $ txInfoOutputs txInfo
+
+  val = maybe mempty (txOutValue . txInInfoResolved) $ findOwnInput ctx
+
   hasBoundedValidity =
     traceIfFalse "hasBoundedValidity check failed" $
       tMax - tMin <= cp
-
-  outValue =
-    maybe mempty (txOutValue . txInInfoResolved) $ findOwnInput ctx
 
   (closedSnapshotNumber, closedUtxoHash, parties', closedContestationDeadline, headId', contesters') =
     -- XXX: fromBuiltinData is super big (and also expensive?)
@@ -351,6 +363,8 @@ checkClose ctx parties initialUtxoHash sig cperiod headPolicyId =
 --   * Add signer to list of contesters.
 --
 --   * No other parameters have changed.
+--
+--   * Value in v_head is preserved
 checkContest ::
   ScriptContext ->
   POSIXTime ->
@@ -370,12 +384,22 @@ checkContest ctx contestationDeadline parties closedSnapshotNumber sig contester
     && mustBeSignedByParticipant ctx headId
     && checkSignedParticipantContestOnlyOnce
     && mustBeWithinContestationPeriod
-    && hasST headId outValue
     && mustUpdateContesters
+    && hasST headId val
     && mustNotChangeParameters
+    && mustPreserveValue
  where
-  outValue =
-    maybe mempty (txOutValue . txInInfoResolved) $ findOwnInput ctx
+  mustPreserveValue =
+    traceIfFalse "head value is not preserved" $
+      -- XXX: Equality on value is very memory intensive as it's defined on
+      -- associative lists and Map equality is implemented. Instead we should be
+      -- more strict and require EXACTLY the same value and compare using a
+      -- simple fold or even compare the serialised bytes.
+      val == val'
+
+  val' = txOutValue . head $ txInfoOutputs txInfo
+
+  val = maybe mempty (txOutValue . txInInfoResolved) $ findOwnInput ctx
 
   mustBeNewer =
     traceIfFalse "too old snapshot" $